+++ src/bin/pktte_sim_anchor.rs
@@ -81,6 +81,7 @@ fn main() {
             reset_times: vec![],
             cens: vec![],
             occasions: vec![],
+            obs_l2: Vec::new(),
             dose_occasions: vec![],
             fremtype: vec![],
             obs_records: vec![ObsRecord::Event {