+++ src/parser/model_parser.rs
@@ -22898,6 +22898,7 @@ if (WT > 70) {
             reset_times: Vec::new(),
             cens: vec![0, 0],
             occasions: vec![1, 1],
+            obs_l2: Vec::new(),
             dose_occasions: vec![1],
             fremtype: Vec::new(),
             obs_records: vec![],
@@ -22918,6 +22919,7 @@ if (WT > 70) {
             &subject.obs_times,
             &subject.obs_raw_times,
             &subject.occasions,
+            &subject.obs_l2,
             &sigma,
             &model.residual_correlations,
         );
@@ -22965,6 +22967,7 @@ if (WT > 70) {
             reset_times: Vec::new(),
             cens: vec![0, 0],
             occasions: vec![1, 1],
+            obs_l2: Vec::new(),
             dose_occasions: vec![1],
             fremtype: Vec::new(),
             obs_records: vec![],