+++ tests/common/mod.rs
@@ -45,6 +45,7 @@ pub fn subject(
         reset_times: vec![],
         cens: vec![0; n],
         occasions: vec![],
+        obs_l2: Vec::new(),
         dose_occasions: vec![],
         fremtype: vec![],
         obs_records: vec![],